Fix CI on AppVeyor (windos) and Travis-CI (linux, osx)
Notes:
- Dependency on MSVC should be added when it is fixed in conda. See
  https://github.com/conda/conda-build/issues/2263
- Yaff cannot be compiled on Windows with Py 2.7. (It would require a
  lot of C code changes because Py27 is always built with MSVC 2008 on
  Windows, which does not even support C99!)
- Rebuilding font caches for matplotlib is slow on Travis.

+++ yaff/pes/grid.c
@@ -31,7 +31,7 @@
 #include "grid.h"
 
 
-double compute_grid3d(double* center, cell_type *cell, double* egrid, long* shape) {
+double compute_grid3d(double* center, cell_type *cell, double* egrid, size_t* shape) {
     double frac[3], e11, e12, e21, e22;
     long indexes[3];
     int i;
